Computer Vision/Rust_corrosion/requirements.txt
--index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124

ultralytics==8.3.218
flask==3.0.3
pillow
numpy
opencv-python
pyyaml
requests
